+++ src/parser/enrichers/feat/_module.mjs
@@ -33,6 +33,7 @@ export { default as Sharpshooter } from "./Sharpshooter.mjs";
 export { default as Slasher } from "./Slasher.mjs";
 export { default as SpeedyRecovery } from "./SpeedyRecovery.mjs";
 export { default as SquireOfSolamniaPreciseStrike } from "./SquireOfSolamniaPreciseStrike.mjs";
+export { default as StrikeOfTheGiants } from "./StrikeOfTheGiants.mjs";
 export { default as TavernBrawler } from "./TavernBrawler.mjs";
 export { default as Telekinetic } from "./Telekinetic.mjs";
 export { default as WarCaster } from "./WarCaster.mjs";