+++ src/parser/enrichers/item/_module.mjs
@@ -22,6 +22,7 @@ export { default as DustOfSneezingAndChoking } from "./DustOfSneezingAndChoking.
 export { default as FarRealmShard } from "./FarRealmShard.mjs";
 export { default as FlameTongue } from "./FlameTongue.mjs";
 export { default as FlyingBroomstick } from "./FlyingBroomstick.mjs";
+export { default as GenericLightSource } from "./GenericLightSource.mjs";
 export { default as GnomengardeGrenade } from "./GnomengardeGrenade.mjs";
 export { default as HammerOfThunderbolts } from "./HammerOfThunderbolts.mjs";
 export { default as HealersKit } from "./HealersKit.mjs";